+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+-   [`FormsClient.GetDraftSubmission()`](./docs/forms-client.md#getdraftsubmission)
+
 ## 1.0.0
 
 Initial production release